backend/requirements.txt
# Financial Document Analyzer - Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# LangChain & AI
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# Development
httpx>=0.27.0  # For testing
